GUI: Add scrolling with mousewheel and buttons. The buttons do not work as I like yet, but the mousehweel works perfectly. Scrolling works ok. The initial list has been changed to test scrolling.
Next step is to create a mouse class that is handy to work with, so the button behaviour can be fixed (and it is easier to read state from the mouse, instead of the game class, which is a faulty design).

Of course, the initial lists need to be fixed as well, as is the availability of their list buttons. Also, the building mechanism must be rewritten. After that, code can be removed from the cGame_logic class.

git-svn-id: http://dune2themaker.googlecode.com/svn/trunk@159 52bf4e17-6a1c-0410-83a1-9b5866916151